Refactor maintenance mode code to maintenance.py
