+++ actions.yaml
@@ -43,12 +43,12 @@ enter-maintenance:
       type: boolean
       default: false
       description: |
-        Do not run the steps but print the action plan.
+        Do not run the operations but print the action plan.
     force:
       type: boolean
       default: false
       description: |
-        Force to enter maintenance mode even if the sanity checks fail.
+        Force to run the operations to enter maintenance mode even if there are failed operations.
     set-noout:
       type: boolean
       default: true
@@ -58,17 +58,21 @@ enter-maintenance:
       type: boolean
       default: false
       description: |
-        Optional to stop and disable OSD service on that node. Defaults to keep the OSD service running when entering maintenance mode.
+        Optional to stop and disable OSD service on that node. Defaults to keep the OSD service
+        running when entering maintenance mode.
     check-only:
       type: boolean
       default: false
       description: |
-        Only run the checks for entering maintenance mode.
+        Only run the sanity checks for entering maintenance mode. Do not confuse with `dry-run`
+        option, which only prints the action plan.
     ignore-check:
       type: boolean
       default: false
       description: |
-        Ignore preflight check for entering maintenance mode.
+        Ignore sanity checks and run the operations for entering maintenance mode. Do not confuse
+        with `force` option, which will ignore any error in the operations and continue the
+        remaining operations.
 exit-maintenance:
   description: |
      Bring the given unit out of maintenance mode.
@@ -77,14 +81,15 @@ exit-maintenance:
       type: boolean
       default: false
       description: |
-        Do not run the steps but print the action plan.
+        Do not run the operations but print the action plan.
     check-only:
       type: boolean
       default: false
       description: |
-        Only run the checks for exiting maintenance mode.
+        Only run the sanity checks for exiting maintenance mode. Do not confuse with `dry-run`
+        option, which only prints the action plan.
     ignore-check:
       type: boolean
       default: false
       description: |
-        Ignore preflight check for exiting maintenance mode.
+        Ignore sanity checks and run the operations for exiting maintenance mode.
